+++ .github/CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ## 2024.2.0
 
+1. [A380X/MFD] Fix mandatory and disabled input fields showing amber dashes instead of grey placeholder boxes - @agarwalrahul2702 (Rahul Agarwal)
 1. [ATSU] Add support for BeyondATC and SayIntentions AI as METAR/ATIS sources - @saschl
 1. [A32NX/TCAS] Improve PFD Resolution Advisory graphics, introduce TCAS bus - @lukecologne (luke)
 1. [A380X/FMS] Fix TOW on FUEL&LOAD page showing gross weight instead of takeoff weight in flight - @HendersonTyler